+++ typings/Pluralization.d.ts
@@ -1,5 +1,10 @@
-import { Pluralizer } from "./typing";
+import { Pluralizer, MakePlural } from "./typing";
 import { I18n } from "./I18n";
+export declare function useMakePlural({ pluralizer, includeZero, ordinal, }: {
+    pluralizer: MakePlural;
+    includeZero?: boolean;
+    ordinal?: boolean;
+}): Pluralizer;
 export declare const defaultPluralizer: Pluralizer;
 export declare class Pluralization {
     private i18n;