+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Prefix your message with one of the following:
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## v4.2.3 - Mar 5, 2023
 
 - [Added] Introduce `useMakePlural(options)` function as way of creating
   pluralizers on top of [make-plural](https://github.com/eemeli/make-plural/).